fix_bugs_hilt_commit

+++ app/src/main/java/com/hamiddev/model/WeatherResponse.kt
@@ -4,8 +4,8 @@ data class WeatherResponse(
     val current: Current,
     val daily: List<Daily>,
     val hourly: List<Hourly>,
-    val lat: Int,
-    val lon: Int,
+    val lat: Double,
+    val lon: Double,
     val timezone: String,
     val timezone_offset: Int
 )
